refactor: 컴포넌트화
