+++ src/config/index.ts
@@ -1,7 +1,5 @@
 import dotenv from 'dotenv';
 
-dotenv.config();
-
 const config = {
   port: process.env.PORT,
   mongo: process.env.MONGO_URL || '',
