+++ src/config/sequelize.ts
@@ -1,7 +1,5 @@
 import dotenv from 'dotenv';
 
-dotenv.config();
-
 export default {
   development: {
     user: process.env.MYSQL_USER,
