--- src/types/user.ts
@@ -1,16 +1,0 @@
-import Todo from './todo';
-
-interface User {
-  id?: string;
-
-  name: string;
-  email: string;
-  password?: string;
-
-  todos?: Todo[];
-
-  thumbnailImage?: string;
-  profileImage?: string;
-}
-
-export default User;
